bindgen
Module callbacks

A public API for more fine-grained customization of bindgen behavior.

Structs

DeriveInfo

Relevant information about a type to which new derive attributes will be added using ParseCallbacks::add_derives.

Enums

DeriveTrait

Which trait to consider when doing the CannotDerive analysis.

EnumVariantCustomBehavior

An enum representing custom handling that can be given to a variant.

EnumVariantValue

A constant value assigned to an enumeration variant.

ImplementsTrait

Whether it is possible or not to automatically derive trait for an item.

IntKind

Which integral type are we dealing with?

MacroParsingBehavior

An enum to allow ignoring parsing of macros.

Traits

ParseCallbacks

A trait to allow configuring different kinds of types in different situations.