Crate bindgen

Generate Rust bindings for C and C++ libraries.

Provide a C/C++ header file, receive Rust FFI code to call into C/C++ functions and use types defined in the header.

See the Builder struct for usage.

Modules

callbacks
A public API for more fine-grained customization of bindgen behavior.

Macros

extra_assert

extra_assert_eq

Structs

Bindings
Generated Rust bindings.

Builder
Configure and generate Rust bindings for a C/C++ header.

ClangVersion
Extracted Clang version data

CodegenConfig
A type used to indicate which kind of items do we have to generate.

Enums

LinkType
The linking type to use with a given library.

RustTarget
Represents the version of the Rust language to target.

Constants

LATEST_STABLE_RUST
Latest stable release of Rust

Statics

RUST_TARGET_STRINGS
Strings of allowed RustTarget values

Functions

builder
Construct a new Builder.

clang_version
Get the major and the minor semvar numbers of Clang's version