bindgen

Module callbacks

A public API for more fine-grained customization of bindgen behavior.

Structs

AttributeInfo

Relevant information about a type to which new attributes will be added using ParseCallbacks::add_attributes.

DeriveInfo

Relevant information about a type to which new derive attributes will be added using ParseCallbacks::add_derives.

DiscoveredItemId

An identifier for a discovered item. Used to identify an aliased type (see DiscoveredItem::Alias)

FieldInfo

Relevant information about a field for which visibility can be determined using ParseCallbacks::field_visibility.

ItemInfo

A struct providing information about the item being passed to ParseCallbacks::generated_name_override.

Enums

DeriveTrait

Which trait to consider when doing the CannotDerive analysis.

DiscoveredItem

Struct passed to ParseCallbacks::new_item_found containing information about discovered items (struct, union, and alias)

EnumVariantCustomBehavior

An enum representing custom handling that can be given to a variant.

EnumVariantValue

A constant value assigned to an enumeration variant.

ImplementsTrait

Whether it is possible or not to automatically derive trait for an item.

IntKind

Which integral type are we dealing with?

ItemKind

An enum indicating the kind of item for an ItemInfo.

MacroParsingBehavior

An enum to allow ignoring parsing of macros.

TypeKind

The kind of the current type.

Traits

ParseCallbacks

A trait to allow configuring different kinds of types in different situations.