Enum bindgen::callbacks::IntKind

pub enum IntKind {
    Bool,
    SChar,
    UChar,
    WChar,
    Char {
        is_signed: bool,
    },
    Short,
    UShort,
    Int,
    UInt,
    Long,
    ULong,
    LongLong,
    ULongLong,
    I8,
    U8,
    I16,
    U16,
    I32,
    U32,
    I64,
    U64,
    I128,
    U128,
    Custom {
        name: &'static str,
        is_signed: bool,
    },
}

Which integral type are we dealing with?

Variants
Bool

A bool.

SChar

A signed char.

UChar

An unsigned char.

WChar

An wchar_t.

Char

A platform-dependent char type, with the signedness support.

Fields of Char

is_signed: bool

Whether the char is signed for the target platform.

Short

A short.

UShort

An unsigned short.

Int

An int.

UInt

An unsigned int.

Long

A long.

ULong

An unsigned long.

LongLong

A long long.

ULongLong

An unsigned long long.

I8

A 8-bit signed integer.

U8

A 8-bit unsigned integer.

I16

A 16-bit signed integer.

U16

Either a char16_t or a wchar_t.

I32

A 32-bit signed integer.

U32

A 32-bit unsigned integer.

I64

A 64-bit signed integer.

U64

A 64-bit unsigned integer.

I128

An int128_t

U128

A uint128_t.

Custom

A custom integer type, used to allow custom macro types depending on range.

Fields of Custom

name: &'static str

The name of the type, which would be used without modification.

is_signed: bool

Whether the type is signed or not.

Methods

impl IntKind

pub fn is_signed(&self) -> bool

Is this integral type signed?

pub fn known_size(&self) -> Option<usize>

If this type has a known size, return it (in bytes). This is to alleviate libclang sometimes not giving us a layout (like in the case when an enum is defined inside a class with template parameters).

pub fn signedness_matches(&self, val: i64) -> bool

Whether this type's signedness matches the value.