bindgen
Enum RustTarget

pub enum RustTarget {

    Nightly,
    Stable_1_77,
    Stable_1_73,
    Stable_1_71,
    Stable_1_68,
    Stable_1_64,
    Stable_1_59,
    Stable_1_47,
    Stable_1_43,
    Stable_1_40,
    Stable_1_36,
    Stable_1_33,
    Stable_1_30,
    Stable_1_28,
    Stable_1_27,
    Stable_1_26,
    Stable_1_25,
    Stable_1_21,
    Stable_1_20,
    Stable_1_19,
    Stable_1_17,
    Stable_1_0,

}

Represents the version of the Rust language to target.

To support a beta release, use the corresponding stable release.

This enum will have more variants added as necessary.

Variants

Nightly

Rust Nightly

vectorcall_abi

ptr_metadata

layout_for_ptr

Stable_1_77

Rust 1.77

offset_of

Stable_1_73

Rust 1.73

thiscall_abi

Stable_1_71

Rust 1.71

c_unwind_abi

Stable_1_68

Rust 1.68

abi_efiapi

Stable_1_64

Rust 1.64

core_ffi_c

Stable_1_59

Rust 1.59

const_cstr

Stable_1_47

Rust 1.47

larger_arrays

Stable_1_43

Rust 1.43

associated_constants

Stable_1_40

Rust 1.40

non_exhaustive

Stable_1_36

Rust 1.36

maybe_uninit

Stable_1_33

Rust 1.33

repr_packed_n

Stable_1_30

๐Deprecated

Rust 1.30

core_ffi_c_void

min_const_fn

Stable_1_28

๐Deprecated

Rust 1.28

repr_transparent

Stable_1_27

๐Deprecated

Rust 1.27

must_use_function

Stable_1_26

๐Deprecated

Rust 1.26

i128_and_u128

Stable_1_25

๐Deprecated

Rust 1.25

repr_align

Stable_1_21

๐Deprecated

Rust 1.21

builtin_clone_impls

Stable_1_20

๐Deprecated

Rust 1.20

associated_const

Stable_1_19

๐Deprecated

Rust 1.19

untagged_union

Stable_1_17

๐Deprecated

Rust 1.17

static_lifetime_elision

Stable_1_0

๐Deprecated

Rust 1.0