Enum bindgen::RustTarget

pub enum RustTarget {

    Stable_1_0,
    Stable_1_17,
    Stable_1_19,
    Stable_1_20,
    Stable_1_21,
    Stable_1_25,
    Stable_1_26,
    Stable_1_27,
    Stable_1_28,
    Stable_1_30,
    Stable_1_33,
    Stable_1_36,
    Stable_1_40,
    Stable_1_47,
    Nightly,

}

Represents the version of the Rust language to target.

To support a beta release, use the corresponding stable release.

This enum will have more variants added as necessary.

Variants

Stable_1_0

Rust stable 1.0

Stable_1_17

Rust stable 1.17

Static lifetime elision (RFC 1623)

Stable_1_19

Rust stable 1.19

Untagged unions (RFC 1444)

Stable_1_20

Rust stable 1.20

Associated constants (PR)

Stable_1_21

Rust stable 1.21

Builtin impls for Clone (PR)

Stable_1_25

Rust stable 1.25

repr(align) (PR)

Stable_1_26

Rust stable 1.26

i128 / u128 support

Stable_1_27

Rust stable 1.27

must_use attribute on functions (PR)

Stable_1_28

Rust stable 1.28

repr(transparent) (PR)

Stable_1_30

Rust stable 1.30

const fn support for limited cases (PR

c_void available in core

Stable_1_33

Rust stable 1.33

repr(packed(N)) (PR)

Stable_1_36

Rust stable 1.36

MaybeUninit instead of mem::uninitialized() (PR)

Stable_1_40

Rust stable 1.40

non_exhaustive enums/structs (Tracking issue)

Stable_1_47

Rust stable 1.47

larger_arrays (Tracking issue)

Nightly

Nightly rust

thiscall calling convention (Tracking issue)

vectorcall calling convention (no tracking issue)