Struct bindgen::CargoCallbacks

pub struct CargoCallbacks;

A ParseCallbacks implementation that will act on file includes by echoing a rerun-if-changed line

When running in side a build.rs script, this can be used to make cargo invalidate the generated bindings whenever any of the files included from the header change:

use bindgen::builder;
let bindings = builder()
    .header("path/to/input/header")
    .parse_callbacks(Box::new(bindgen::CargoCallbacks))
    .generate();