bindgen

Struct CargoCallbacks

pub struct CargoCallbacks { /* private fields */ }

A ParseCallbacks implementation that will act on file includes by echoing a rerun-if-changed line and on env variable usage by echoing a rerun-if-env-changed line

When running inside a build.rs script, this can be used to make cargo invalidate the generated bindings whenever any of the files included from the header change:

use bindgen::builder;
let bindings = builder()
    .header("path/to/input/header")
    .parse_callbacks(Box::new(bindgen::CargoCallbacks::new()))
    .generate();

Implementations

impl CargoCallbacks

pub fn new() -> Self

Create a new CargoCallbacks value.

pub fn rerun_on_header_files(self, doit: bool) -> Self

Whether Cargo should re-run the build script if any of the input header files has changed.

This option is enabled by default unless the deprecated CargoCallbacks constructor is used.